requirements/test.txt
-r base.txt

coverage==4.4.1
mock==2.0.0
factory-boy==2.8.1
faker==0.7.15
pytest==3.1.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-mock==1.6.0
pytest-pep8==1.0.6
pytest-flakes==2.0.0
